pyproject.toml
[project]
name = "z2p"
version = "0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.0",
    "fastapi>=0.119.0",
    "granian>=2.5.5",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "stamina>=25.1.0",
]

[dependency-groups]
dev = [
    "coverage>=7.10.7",
    "fastapi[standard]>=0.119.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
branch = true
parallel = true
source = ["z2p_svc", "tests"]

[tool.coverage.report]
show_missing = true
skip_covered = true
omit = ["src/**/asgi.py"]

[tool.pytest.ini_options]
# 修复 Windows 上 colorama 导致的 I/O 错误
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
# 异步测试配置
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# 测试发现
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# 标记
markers = [
    "asyncio: mark test as async",
    "unit: 单元测试",
    "integration: 集成测试",
    "system: 系统测试",
    "e2e: 端到端测试",
    "slow: 慢速测试",
    "network: 需要网络连接的测试",
]
# Windows 特定配置：禁用颜色输出以避免 colorama 问题
norecursedirs = [".git", ".tox", "dist", "build", "*.egg"]
